modularize scrutineer: extract scrutineer command line module - extract the existing ES7+Jdbc streams + commandline into submodule which I still named artifactId as "scrutineer", it provide backward-compatibility for clients still relying on "scrutineer jar" - this "scrutineer" module can build rpm/distribution tarball - rename artifactId of the root pom from "scrutineer" to "scrutineer-pom"
