+++ elasticsearch7/pom.xml
@@ -3,13 +3,14 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>scrutineer</artifactId>
+        <artifactId>scrutineer-pom</artifactId>
         <groupId>com.aconex.scrutineer</groupId>
         <version>7.9.4-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>elasticsearch7</artifactId>
+    <name>Elastic Search(7) Stream</name>
 
     <properties>
         <httpclient.version>4.5.13</httpclient.version>
