+++ scrutineer/src/test/java/com/aconex/scrutineer/CmdLineOptionsSSLVerificationModeValidatorTest.java
