+++ scrutineer/src/test/java/com/aconex/scrutineer/functional/ScrutineerIntegrationTest.java
