+++ scrutineer/src/test/java/com/aconex/scrutineer/functional/ScrutineerNumericIntegrationTest.java
