+++ scrutineer/src/test/resources/com/aconex/scrutineer/functional/es-bulkindex.json
