+++ scrutineer/src/test/resources/com/aconex/scrutineer/functional/es-numericbulkindex.json
