+++ scrutineer/src/test/resources/com/aconex/scrutineer/functional/fullintegrationtest.xml
