+++ scrutineer/src/test/resources/com/aconex/scrutineer/functional/fullnumericintegrationtest.xml
