--- src/main/java/com/aconex/scrutineer/javautil/SystemTimeSource.java
@@ -1,9 +1,0 @@
-package com.aconex.scrutineer.javautil;
-
-public class SystemTimeSource implements TimeSource {
-
-    @Override
-    public long getCurrentTime() {
-        return System.currentTimeMillis();
-    }
-}
