--- src/main/java/com/aconex/scrutineer/javautil/TimeSource.java
@@ -1,7 +1,0 @@
-package com.aconex.scrutineer.javautil;
-
-public interface TimeSource {
-
-    long getCurrentTime();
-
-}
