+++ verifier/src/main/java/com/aconex/scrutineer/CoincidentFilteredStreamVerifierListener.java
