+++ verifier/src/main/java/com/aconex/scrutineer/CompositeStreamVerifierListener.java
