+++ verifier/src/main/java/com/aconex/scrutineer/DeletionVerifier.java
