+++ verifier/src/test/java/com/aconex/scrutineer/CompositeStreamVerifierListenerTest.java
