+++ verifier/src/test/java/com/aconex/scrutineer/PrintStreamOutputVersionStreamVerifierListenerTest.java
